backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic-settings
python-jose[cryptography]
authlib
google-generativeai
pydantic-ai
redis
httpx
pytest
hypothesis
pytest-asyncio
fakeredis
pytest-cov
